Version 3.0
Updated to handle files with comma delimited data, and .txt files with missing time data. Time is being calculated as a function of FPS (FPS setting is implemented as a GUI feature) and calculated time values, along with the original data are being saved inside the default directory without overwriting the original .txt file.